+++ src/main/java/seedu/address/commons/events/ui/PersonPanelSelectionChangedEvent.java
@@ -1,7 +1,7 @@
 package seedu.address.commons.events.ui;
 
 import seedu.address.commons.events.BaseEvent;
-import seedu.address.model.person.ReadOnlyTask;
+import seedu.address.model.task.ReadOnlyTask;
 
 /**
  * Represents a selection change in the Person List Panel
