+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -3,8 +3,8 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.person.ReadOnlyTask;
-import seedu.address.model.person.UniquePersonList.TaskNotFoundException;
+import seedu.address.model.task.ReadOnlyTask;
+import seedu.address.model.task.UniquePersonList.TaskNotFoundException;
 
 /**
  * Deletes a task identified using it's last displayed index from the task list.
