+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -2,8 +2,8 @@ package seedu.address.model;
 
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.ReadOnlyTask;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.task.ReadOnlyTask;
 
 /**
  * Unmodifiable view of an address book
