+++ src/main/java/seedu/address/model/task/ReadOnlyTask.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.task;
 
 import seedu.address.model.tag.UniqueTagList;
