+++ src/main/java/seedu/address/model/task/UniquePersonList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.task;
 
 import java.util.Iterator;
 import java.util.List;
