+++ src/test/java/guitests/SampleDataTest.java
@@ -6,7 +6,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import seedu.address.model.AddressBook;
-import seedu.address.model.person.Task;
+import seedu.address.model.task.Task;
 import seedu.address.model.util.SampleDataUtil;
 import seedu.address.testutil.TestUtil;
 @Ignore
