+++ src/test/java/guitests/SelectCommandTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import seedu.address.model.person.ReadOnlyTask;
+import seedu.address.model.task.ReadOnlyTask;
 @Ignore
 public class SelectCommandTest extends AddressBookGuiTest {
