+++ src/test/java/seedu/address/model/person/PhoneTest.java
@@ -5,6 +5,8 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.address.model.task.Phone;
+
 public class PhoneTest {
 
     @Test
