+++ src/test/java/seedu/address/testutil/TestTask.java
@@ -1,8 +1,8 @@
 package seedu.address.testutil;
 
-import seedu.address.model.person.Name;
-import seedu.address.model.person.ReadOnlyTask;
 import seedu.address.model.tag.UniqueTagList;
+import seedu.address.model.task.Name;
+import seedu.address.model.task.ReadOnlyTask;
 
 /**
  * A mutable person object. For testing only.
